Type the announcer through the section event map and surface the trade
