+++ test/util/fire-event.test.ts
@@ -1,9 +1,5 @@
 import { describe, expect, it } from "vitest";
-import {
-  fireEvent,
-  fireFromAnchor,
-  prepareYamlUpdated,
-} from "../../src/util/fire-event.js";
+import { fireEvent, fireFromAnchor } from "../../src/util/fire-event.js";
 
 describe("fireEvent", () => {
   it("dispatches a bubbling composed CustomEvent with the detail", () => {
@@ -69,26 +65,5 @@ describe("fireFromAnchor", () => {
     fireFromAnchor(host, false, null, "yaml-updated", { yaml: "x" });
 
     expect(seen).toEqual([]);
-  });
-});
-
-describe("prepareYamlUpdated", () => {
-  it("captures the anchor before the awaits and carries the basis", () => {
-    const anchor = new EventTarget();
-    const host = Object.assign(new EventTarget(), {
-      parentNode: anchor as unknown as ParentNode,
-    });
-    const seen: { yaml: string; basedOn?: string }[] = [];
-    anchor.addEventListener("yaml-updated", (e) =>
-      seen.push((e as CustomEvent<{ yaml: string; basedOn?: string }>).detail)
-    );
-
-    const announce = prepareYamlUpdated(host);
-    // The host unmounts mid round trip; the announcement still rides
-    // the anchor captured up front and carries the write's basis.
-    host.parentNode = null as unknown as ParentNode;
-    announce(false, { yaml: "b:\n", basedOn: "a:\n" });
-
-    expect(seen).toEqual([{ yaml: "b:\n", basedOn: "a:\n" }]);
   });
 });
